+++ src/cqindustry/chip/__init__.py
@@ -5,3 +5,4 @@ from .GuardRail import GuardRail
 from .make_platform_rails import make_platform_rails
 from .ChipCan import ChipCan
 from .ChipTower import ChipTower
+from .Conduit import Conduit